feat(LottoResultView): 발행한 로또의 내역을 출력하는 뷰 클래스를 분리
