+++ src/main/java/lotto/view/constants/SymbolType.java
@@ -1,10 +1,7 @@
 package lotto.view.constants;
 
 public enum SymbolType {
-    OUTPUT_SEPARATOR(", "),
     INPUT_SEPARATOR(","),
-    PREFIX("["),
-    POSTFIX("]"),
     NEWLINE("\n"),
     WINNING_RESULT_SEPARATOR(" - ");
